requirements.txt
langchain==0.3.1
langchain_groq==0.2.0
python-dotenv==1.0.1
streamlit==1.35.0
